requirements.txt
opencv-python==4.12.0.88
numpy==2.2.2
scipy==1.15.1
pandas==2.2.3
streamlit>=1.28.0
pillow>=10.0.0